requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.3.0
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coverage==7.6.9
dill==0.4.0
distlib==0.4.0
execnet==2.1.1
Faker==37.6.0
filelock==3.19.1
greenlet==3.2.4
identify==2.6.14
idna==3.10
imagesize==1.4.1
iniconfig==2.1.0
isort==5.13.2
keyring==25.6.0
markdown-it-py==4.0.0
more-itertools==10.8.0
mypy_extensions==1.1.0
mysql-connector-python
nh3==0.3.0
nodeenv==1.9.1
numpy==2.2.6
openai>=1.0.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
platformdirs==4.4.0
playwright==1.55.0
pluggy==1.6.0
pre_commit==4.0.1
pyee==13.0.0
pyproject_hooks==1.2.0
pytest==8.2.0
pytest-base-url==2.1.0
pytest-cov==6.0.0
pytest-dependency==0.6.0
pytest-html==4.1.1
pytest_html_merger==0.1.0
pytest-jira-xray==0.9.2
pytest-json-report==1.5.0
pytest-metadata==3.1.1
pytest-playwright==0.7.1
pytest-rerunfailures==13.0
pytest-xdist==3.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
virtualenv==20.34.0
